requirements.txt

pandas==1.0.5
numpy==1.19.4
seaborn==0.10.1
torch==1.4.0
Keras==2.4.3
opencv_python==4.4.0.46
scipy==1.5.0
matplotlib==3.2.2
tensorflow==2.4.0
tqdm==4.47.0
efficientnet==1.1.1
ipython==7.27.0
librosa==0.8.1
Pillow==8.3.2
scikit_learn==0.24.2
